requirements.txt
alembic==1.5.6
aniso8601==9.0.1
astroid==2.5.1
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
Faker==6.6.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-expects-json==1.5.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
flask-restplus==0.13.0
flask-restx==0.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
idna==2.10
iniconfig==1.1.1
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.3
jsonschema==3.2.0
kanpai==0.1.15
lazy-object-proxy==1.5.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.16
psycopg2==2.8.6
ptyprocess==0.7.0
pudb==2020.1
py==1.10.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pylint==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.3
urwid==2.1.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
